+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+-------------------
+* Always auto-resubscribe all subs after reconnect
+
 v1.0.0 (2017-03-07)
 -------------------
 * Initial release